+++ framework/model/src/test/java/is/codion/framework/model/DefaultEntityConditionModelTest.java
@@ -79,7 +79,7 @@ public class DefaultEntityConditionModelTest {
 		//no search columns defined for master entity
 		ForeignKeyConditionModel masterModel = model.get(Detail.MASTER_FK);
 		assertThrows(IllegalStateException.class, () ->
-						masterModel.equalSearchModel().search().result());
+						masterModel.equalSearchModel().search().perform());
 	}
 
 	@Test