+++ changelog.md
@@ -19,6 +19,7 @@ Codion Change Log
 - EntityEditor.DetailEditors.remove() added.
 - EntitySearchModel.Builder.searchColumns() overloaded with varargs parameter.
 - EntitySearchModel.Builder.searchColumns() renamed search(columns).
+- EntitySearchModel.Search.result() renamed perform().
 ### is.codion.swing.framework.ui
 - EditorComponents.entries() and editorComponent() bug fixed, now exclude entries without a component.
 - EditorComponents.detail() added, removal implemented.